third_party/blink/web_tests/external/wpt/html/cross-origin-opener-policy/historical/popup-same-origin-unsafe-allow-outgoing-with-cross-origin.https-expected.txt

This is a testharness.js-based test.
PASS same-origin unsafe-allow-outgoing document opening popup to https://not-web-platform.test:8444 with COOP: ""
PASS same-origin unsafe-allow-outgoing document opening popup to https://not-web-platform.test:8444 with COOP: "unsafe-none"
PASS same-origin unsafe-allow-outgoing document opening popup to https://not-web-platform.test:8444 with COOP: "jibberish"
PASS same-origin unsafe-allow-outgoing document opening popup to https://not-web-platform.test:8444 with COOP: "same-site"
PASS same-origin unsafe-allow-outgoing document opening popup to https://not-web-platform.test:8444 with COOP: "same-site unsafe-allow-outgoing"
PASS same-origin unsafe-allow-outgoing document opening popup to https://not-web-platform.test:8444 with COOP: "same-origin unsafe-allow-outgoing"
FAIL same-origin unsafe-allow-outgoing document opening popup to https://not-web-platform.test:8444 with COOP: "same-origin" assert_equals: name expected "" but got "same-origin unsafe-allow-outgoing_to_CROSS_ORIGIN_same-origin"
FAIL same-origin unsafe-allow-outgoing document opening popup to https://not-web-platform.test:8444 with COOP: "same-origin-allow-popups" assert_equals: name expected "" but got "same-origin unsafe-allow-outgoing_to_CROSS_ORIGIN_same-origin-allow-popups"
Harness: the test ran to completion.
